spreadsheet editor using stage model first pass
